Refactor GitHub Actions workflow by renaming the quality job to lint for clarity. Update layout imports to use absolute paths for Navbar and Footer components. Add new Navbar and Footer components along with a NavbarSidebar for improved navigation structure.
