+++ .github/workflows/pull-request.yaml
@@ -5,7 +5,7 @@ on:
     branches: [main]
 
 jobs:
-  quality:
+  lint:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout code
